Renamed this from util to htmlComponent to match the component inside it
Thought about moving the module to lib/utils, but I think it makes sense to keep it here. The utils name doesn't quite make sense since we never ended up making any other utils
